Update dependencies + gradle
Also removed fabric api, as it is unused, and changed fabric language kotlin from an included dependency, to an external dependency as it is quite large. Also removed fabric-api-base as an included dependency, and added fabric-api as an external dependency. Fabric-api is a common dependency so it is most likely already installed.